pyproject.toml
[project]
name = "shipment-tracker-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "pydantic>=2.11.5",
    "hypercorn>=0.17.0",
    "pandas>=2.3.0",
    "pytest>=8.4.0",
    "pytest-cov>=6.1.1",
    "python-dotenv>=1.1.0",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "httpx>=0.27.0",
]

[project.scripts]
shipment-api = "shipment_tracker_api.main:main"
